Adds public IP and Runpod TCP port to state
Adds the public IP address and Runpod TCP port 22 to the instance and network state information.

This enables external access and better monitoring of the environment.

Also, improves rsync command generation to utilize public IP and port for Runpod environments.
Two new database migrations are added for the new fields.